health_data_aggregator.py
import logging
from decouple import config

# Add other necessary imports here

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

API_KEY = config("API_KEY", default="YourDefaultAPIKey")


def main():
    logger.info("Starting the Health Data Aggregator...")
    # Your core application logic here


if __name__ == "__main__":
    main()
