requirements.txt
python-decouple
Flask
pytest